requirements.txt
fastapi==0.115.5
python-multipart==0.0.9
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
requests==2.32.3
celery==5.3.6
redis==5.0.1
boto3==1.34.145


# LangChain/LangGraph stack
langchain==1.0.11
langchain-core==1.0.11
langchain-community==1.0.11
langchain-text-splitters==0.3.2
langchain-openai==1.0.6
langgraph==0.3.1
langgraph-checkpoint-sqlite==3.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
prometheus-client==0.20.0
pdfplumber==0.11.3
pillow==11.0.0
pytesseract==0.3.13
pydantic-settings==2.6.0
